Fix broken relationships when using FEM and CoreData
We ran into this crash: "Unacceptable type of value for to-many relationship: property = "foo"; desired type = NSSet; given type = __NSArrayM; value = (...)"
Turns out, that class_getProperty does not work for NSManagedObject and returns NULL.
In this case fem_propertyRepresentation defaults to [NSArray class] which causes the
above error.

credits to @dlalic for pointing this out and providing a fix.